cleaning more uses of dispatch from evalframe
